chore: move postbox to provider
