+++ src/extension/content-script/index.tsx
@@ -1,5 +1,4 @@
 import './injections/Welcome' // ? Inject welcome
-import './injections/PostBox' // ? Inject postbox
 import './injections/Posts' // ? Inject all posts
 import './injections/ProfilePage' // ? Inject to ProfilePage
 import './tasks' // ? AutomatedTabTask Run tasks when invoked by background page
