+++ src/social-network-provider/facebook.com/UI/injectPostBox.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { LiveSelector, MutationObserverWatcher } from '@holoflows/kit'
-import { AdditionalPostBox } from '../../../components/InjectedComponents/AdditionalPostBox'
 import { renderInShadowRoot } from '../../../utils/jss/renderInShadowRoot'
+import { AdditionalPostBox } from '../../../components/InjectedComponents/AdditionalPostBox'
 
 let composeBox: LiveSelector<Element>
 if (location.hostname.match('m.facebook.com')) {
@@ -12,8 +12,10 @@ if (location.hostname.match('m.facebook.com')) {
         .map(x => x.lastElementChild)
         .map(x => x.firstElementChild)
 }
-const watcher = new MutationObserverWatcher(composeBox)
-    .enableSingleMode()
-    .setDomProxyOption({ afterShadowRootInit: { mode: 'closed' } })
-    .startWatch()
-renderInShadowRoot(<AdditionalPostBox />, watcher.firstVirtualNode.afterShadow)
+export function injectPostBoxFacebook() {
+    const watcher = new MutationObserverWatcher(composeBox)
+        .enableSingleMode()
+        .setDomProxyOption({ afterShadowRootInit: { mode: 'closed' } })
+        .startWatch()
+    renderInShadowRoot(<AdditionalPostBox />, watcher.firstVirtualNode.afterShadow)
+}
